pyproject.toml
[project]
name = "comment-filter-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "reflex==0.7.14",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.11.12",
]
